+++ forgerock-auth/src/main/java/org/forgerock/android/auth/FRUser.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019 - 2021 ForgeRock. All rights reserved.
+ * Copyright (c) 2019 - 2022 ForgeRock. All rights reserved.
  *
  * This software may be modified and distributed under the terms
  * of the MIT license. See the LICENSE file for details.
@@ -333,7 +333,7 @@ public class FRUser {
                 }
             };
 
-            AppAuthFragment.init(manager, this);
+            AppAuthFragment.launch(manager, this);
 
         }